[+] coding style in my_rpg.c

+++ src/my_rpg.c
@@ -9,42 +9,20 @@
 #include "struct.h"
 #include "enum.h"
 
-void game_change(game_t *game)
+void draw_scene2(game_scene_t scene, game_setting_t *settings, game_stat_t
+*stats)
 {
-    set_scale(game);
-    set_cursor_pos(game->scenes[game->settings->current], game->settings);
-    for (int tmp = 0; tmp < game->scenes[game->settings->current].how_many[1];
-    tmp++)
-        if (game->scenes[game->settings->current].sounds[tmp]._activated ==
-        true) {
-            set_music(game);
-            stop_desactivated_music(game);
-        }
-    dick_move_to_spawn(game);
-    particules_star(game);
-    teleport_player_boss(game);
-    if (game->settings->current == BOSS)
-        change_view_boss(game->settings);
-    if (game->settings->current == MAIN_MENU)
-        change_main_menu(game->settings, game->scenes, game->stats);
-    if (game->settings->current >= TOWN && game->settings->current <= CAMP)
-        change_maps(game);
-    if (game->settings->current >= VICTORY && game->settings->current <= DEFEAT)
-        change_vic_def(game);
-    clock(game->scenes, game->settings);
-    if (game->settings->current == TOWN) {
-        particules_move_green(game);
-        particules_move_pink(game);
-        particules_move_orange(game);
-    }
-    if (game->settings->_paused == false) {
-        player_interaction(game);
-        enemies_detect_player(game);
-        enemies_detect_hit(game);
-        enemies_attack_player(game);
-        loop_rect_enemies(game);
-    }
-    player_rect_move(game->scenes, game->stats, game->settings);
+    for (int tmp = 0; tmp < scene.how_many[2]; tmp++)
+        if (scene.buttons[tmp].state >= 0)
+            sfRenderWindow_drawRectangleShape(settings->window, scene
+            .buttons[tmp].shape, NULL);
+    for (int tmp = 0; tmp < scene.how_many[3]; tmp++)
+        if (scene.texts[tmp].state >= 0)
+            sfRenderWindow_drawText(settings->window, scene.texts[tmp].text,
+            NULL);
+    if (settings->current == MAIN_MENU && stats->player.speed > 0)
+        sfRenderWindow_drawSprite(settings->window, stats->player.sprite, NULL);
+    draw_cursor(scene, settings);
 }
 
 void draw_scene(game_scene_t scene, game_setting_t *settings, game_stat_t
@@ -60,30 +38,11 @@ void draw_scene(game_scene_t scene, game_setting_t *settings, game_stat_t
         if (scene.objs[tmp].speed >= 1)
             sfRenderWindow_drawSprite(settings->window, scene.objs[tmp]
             .sprite, NULL);
-    for (int tmp = 0; tmp < scene.how_many[2]; tmp++)
-        if (scene.buttons[tmp].state >= 0)
-            sfRenderWindow_drawRectangleShape(settings->window, scene
-            .buttons[tmp].shape, NULL);
-    for (int tmp = 0; tmp < scene.how_many[3]; tmp++)
-        if (scene.texts[tmp].state >= 0)
-            sfRenderWindow_drawText(settings->window, scene.texts[tmp].text,
-            NULL);
-    if (settings->current == MAIN_MENU && stats->player.speed > 0)
-        sfRenderWindow_drawSprite(settings->window, stats->player.sprite, NULL);
-    draw_cursor(scene, settings);
+    draw_scene2(scene, settings, stats);
 }
 
-int my_rpg(void)
+void game_loop(game_t *game)
 {
-    game_t *game = malloc(sizeof(game_t));
-
-    if (game == NULL)
-        return 84;
-    game = create_game(game);
-    if (game == NULL)
-        return 84;
-    if (assets_load(game) == 84)
-        return 84;
     while (sfRenderWindow_isOpen(game->settings->window)) {
         sfRenderWindow_clear(game->settings->window, sfBlack);
         game_change(game);
@@ -99,6 +58,20 @@ int my_rpg(void)
             analyse_events(game);
         sfRenderWindow_display(game->settings->window);
     }
+}
+
+int my_rpg(void)
+{
+    game_t *game = malloc(sizeof(game_t));
+
+    if (game == NULL)
+        return 84;
+    game = create_game(game);
+    if (game == NULL)
+        return 84;
+    if (assets_load(game) == 84)
+        return 84;
+    game_loop(game);
     destroy_all(game);
     return (0);
 }
